Added the heliospectraSTSdaemon to the SeaBreeze project. Given that there are no proprietary concepts or information, and since the daemon is completely reliant upon SeaBreeze to operate, it seemed appropriate. The daemon has been tested on OS X and raspberry pi. There appears to be some kind of USB bug with the STS when OS X sleeps, causing the daemon to misbehave, which has not yet been identified.
A GUI for the daemon was written using the rapid development environment Xojo. A zip file has been included in the build (which may be removed later if wanted) with executables for Windows, Linux and OS X. Xojo compiles for all three targets. The source for the Xojo project was included in XML so that it could be versioned.